# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_alemancenter_fiber-api_internal_models.Conversation:
    properties:
      created_at:
        type: string
      id:
        type: integer
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user1:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.User'
      user1_id:
        type: integer
      user2:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.User'
      user2_id:
        type: integer
    type: object
...
  github_com_alemancenter_fiber-api_internal_models.Message:
    properties:
      body:
        type: string
      conversation:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.Conversation'
      conversation_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_chat:
        type: boolean
      is_deleted:
        type: boolean
      is_draft:
        type: boolean
      is_important:
        type: boolean
      read:
        type: boolean
      recipient:
        allOf:
        - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.User'
        description: Recipient is populated at query time from the conversation, not
          stored in DB.
      sender:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.User'
      sender_id:
        type: integer
      subject:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_alemancenter_fiber-api_internal_models.Permission:
    properties:
      created_at:
        type: string
      guard_name:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_alemancenter_fiber-api_internal_models.PushToken:
    properties:
      created_at:
        type: string
      id:
        type: integer
      platform:
        type: string
      token:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  github_com_alemancenter_fiber-api_internal_models.Role:
    properties:
      created_at:
        type: string
      guard_name:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.Permission'
        type: array
      updated_at:
        type: string
    type: object
...
  github_com_alemancenter_fiber-api_internal_models.User:
    properties:
      bio:
        type: string
      country:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      gender:
        type: string
      google_id:
        type: string
      id:
        type: integer
      job_title:
        type: string
      last_activity:
        type: string
      last_seen:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.Permission'
        type: array
      phone:
        type: string
      profile_photo_path:
        type: string
      push_tokens:
        items:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.PushToken'
        type: array
      roles:
        description: Relationships
        items:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.Role'
        type: array
      social_links:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_alemancenter_fiber-api_internal_utils.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.PaginationMeta'
      pagination:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.PaginationMeta'
      success:
        type: boolean
      token:
        type: string
    type: object
  github_com_alemancenter_fiber-api_internal_utils.PaginationMeta:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
...
  internal_handlers_messages.SaveDraftRequest:
    properties:
      body:
        type: string
      recipient_id:
        type: integer
      subject:
        type: string
    required:
    - body
    type: object
  internal_handlers_messages.SendMessageRequest:
    properties:
      body:
        minLength: 1
        type: string
      recipient_id:
        type: integer
      subject:
        type: string
    required:
    - body
    - recipient_id
    type: object
...
host: api.alemancenter.com
info:
  contact:
    email: support@alemancenter.com
    name: API Support
  description: Backend API for Alemancenter Educational Platform.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Alemancenter API
  version: 2.0.0
paths:
...
  /dashboard/messages/{id}:
    delete:
      description: Soft-delete a message
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      - FrontendKeyAuth: []
      summary: Delete Message
      tags:
      - Messages
    get:
      description: Returns the details of a single message by ID
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.Message'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      - FrontendKeyAuth: []
      summary: Get Message
      tags:
      - Messages
  /dashboard/messages/{id}/important:
    post:
      description: Toggle the importance flag (star) on a message
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      - FrontendKeyAuth: []
      summary: Toggle Message Importance
      tags:
      - Messages
  /dashboard/messages/{id}/read:
    post:
      description: Mark a specific received message as read
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      - FrontendKeyAuth: []
      summary: Mark Message as Read
      tags:
      - Messages
  /dashboard/messages/draft:
    post:
      consumes:
      - application/json
      description: Save a message as a draft for later sending
      parameters:
      - description: Draft message payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_messages.SaveDraftRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.Message'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      - FrontendKeyAuth: []
      summary: Save Draft Message
      tags:
      - Messages
  /dashboard/messages/drafts:
    get:
      description: Returns a paginated list of draft messages for the authenticated
        user
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.Message'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      - FrontendKeyAuth: []
      summary: Get Draft Messages
      tags:
      - Messages
  /dashboard/messages/inbox:
    get:
      description: Returns a paginated list of received messages for the authenticated
        user
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.Message'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      - FrontendKeyAuth: []
      summary: Get Inbox Messages
      tags:
      - Messages
  /dashboard/messages/send:
    post:
      consumes:
      - application/json
      description: Send a new message to another user
      parameters:
      - description: Message payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_messages.SendMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.Message'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      - FrontendKeyAuth: []
      summary: Send Message
      tags:
      - Messages
  /dashboard/messages/sent:
    get:
      description: Returns a paginated list of messages sent by the authenticated
        user
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.Message'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      - FrontendKeyAuth: []
      summary: Get Sent Messages
      tags:
      - Messages
...
schemes:
- https
- http
security:
- FrontendKeyAuth: []
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  FrontendKeyAuth:
    description: Frontend identifier key for public endpoints.
    in: header
    name: X-Frontend-Key
    type: apiKey
swagger: "2.0"